requirements.txt
APScheduler==3.10.4
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.7
django-apscheduler==0.6.2
django-filter==23.5
djangorestframework==3.14.0
idna==3.6
psycopg2-binary==2.9.9
pytz==2024.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.0
